Added orders fetching

+++ DiscogsClient/Data/Result/DiscogsNewListing.cs
@@ -6,7 +6,7 @@ namespace DiscogsClient.Data.Result
 {
     public class DiscogsNewListing
     {
-        public int listing_id { get; set; }
+        public Int64 listing_id { get; set; }
         public string resource_url { get; set; }
     }
 }
